ov/pyproject.toml
[project]
name = "ov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "asyncmy>=0.2.11",
    "fastapi>=0.135.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
    "cryptography>=46.0.5",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "asyncio>=4.0.0",
    "logging>=0.4.9.6",
    "pathlib>=1.0.1",
    "pyjwt>=2.12.1",
    "pandas[excel]>=3.0.2",
    "asyncpg>=0.31.0",
]
